// otel-transformation/src/lib/telemetry/sample-data.ts
import { ResourceSpan } from '@/types/telemetry-types';

/**
 * Hardcoded OpenTelemetry resource span sample data
 * From: opentelemetry-demo-frontendproxy Kubernetes pod
 * Contains 22 resource attributes, 1 scope, 1 span with 20 attributes
 */
export const SAMPLE_TELEMETRY_DATA: { resourceSpans: ResourceSpan[] } = {
  resourceSpans: [
    {
      resource: {
        attributes: [
          {
            key: 'test.resource.id',
            value: { stringValue: '17955346340932591855' },
          },
          {
            key: 'test.resource.name',
            value: {
              stringValue:
                'opentelemetry-demo-frontendproxy-58b488b55d-g7c4t',
            },
          },
          {
            key: 'test.resource.type',
            value: { stringValue: 'k8s.pod' },
          },
          {
            key: 'app.kubernetes.io/component',
            value: { stringValue: 'frontendproxy' },
          },
          {
            key: 'app.kubernetes.io/instance',
            value: { stringValue: 'opentelemetry-demo' },
          },
          {
            key: 'app.kubernetes.io/name',
            value: { stringValue: 'opentelemetry-demo-frontendproxy' },
          },
          {
            key: 'test.auth.token',
            value: { stringValue: 'hxZyXot' },
          },
          {
            key: 'k8s.deployment.name',
            value: { stringValue: 'opentelemetry-demo-frontendproxy' },
          },
          {
            key: 'k8s.deployment.uid',
            value: { stringValue: '79e597ce-4ece-45cc-8672-401819073dd5' },
          },
          {
            key: 'k8s.namespace.name',
            value: { stringValue: 'otel-demo' },
          },
          {
            key: 'k8s.node.name',
            value: {
              stringValue: 'ip-10-1-47-83.eu-west-1.compute.internal',
            },
          },
          {
            key: 'k8s.pod.ip',
            value: { stringValue: '10.1.47.177' },
          },
          {
            key: 'k8s.pod.name',
            value: {
              stringValue:
                'opentelemetry-demo-frontendproxy-58b488b55d-g7c4t',
            },
          },
          {
            key: 'k8s.pod.start_time',
            value: { stringValue: '2025-08-29T02:18:34Z' },
          },
          {
            key: 'k8s.pod.uid',
            value: { stringValue: '653ac4b8-40f5-4776-b5fe-9047e53daf09' },
          },
          {
            key: 'k8s.replicaset.name',
            value: {
              stringValue: 'opentelemetry-demo-frontendproxy-58b488b55d',
            },
          },
          {
            key: 'k8s.replicaset.uid',
            value: { stringValue: '35fd8bf2-07d2-4282-8409-30239b3e021d' },
          },
          {
            key: 'kubectl.kubernetes.io/restartedAt',
            value: { stringValue: '2025-08-29T12:18:33+10:00' },
          },
          {
            key: 'opentelemetry.io/name',
            value: { stringValue: 'opentelemetry-demo-frontendproxy' },
          },
          {
            key: 'pod-template-hash',
            value: { stringValue: '58b488b55d' },
          },
          {
            key: 'service.name',
            value: { stringValue: 'frontendproxy' },
          },
          {
            key: 'service.namespace',
            value: { stringValue: 'opentelemetry-demo' },
          },
        ],
        droppedAttributesCount: 0,
      },
      schemaUrl: 'https://opentelemetry.io/schemas/1.36.0',
      scopeSpans: [
        {
          scope: {
            name: '',
            version: '',
            attributes: [],
            droppedAttributesCount: 0,
          },
          schemaUrl: 'https://opentelemetry.io/schemas/1.36.0',
          spans: [
            {
              traceId: '7Wix1az2dfew4KtQKYRLFg==',
              spanId: '7HNRdwr51gU=',
              parentSpanId: '',
              name: 'ingress',
              kind: 2, // SpanKind.SERVER
              startTimeUnixNano: '1758885889803164000',
              endTimeUnixNano: '1758885889809946000',
              attributes: [
                {
                  key: 'component',
                  value: { stringValue: 'proxy' },
                },
                {
                  key: 'test.operation.name',
                  value: { stringValue: 'GET /ping' },
                },
                {
                  key: 'test.operation.ruleId',
                  value: { stringValue: 'http with url' },
                },
                {
                  key: 'test.operation.type',
                  value: { stringValue: 'http' },
                },
                {
                  key: 'test.span.type',
                  value: { stringValue: 'http' },
                },
                {
                  key: 'downstream_cluster',
                  value: { stringValue: '-' },
                },
                {
                  key: 'guid:x-request-id',
                  value: { stringValue: '17f5e324-a633-9a19-a06f-a7b8e08fb1a4' },
                },
                {
                  key: 'http.protocol',
                  value: { stringValue: 'HTTP/1.1' },
                },
                {
                  key: 'http.request.method',
                  value: { stringValue: 'GET' },
                },
                {
                  key: 'http.response.status_code',
                  value: { stringValue: '404' },
                },
                {
                  key: 'node_id',
                  value: { stringValue: '' },
                },
                {
                  key: 'peer.address',
                  value: { stringValue: '10.1.217.83' },
                },
                {
                  key: 'request_size',
                  value: { stringValue: '0' },
                },
                {
                  key: 'response_flags',
                  value: { stringValue: '-' },
                },
                {
                  key: 'response_size',
                  value: { stringValue: '4144' },
                },
                {
                  key: 'upstream_cluster',
                  value: { stringValue: 'frontend' },
                },
                {
                  key: 'upstream_cluster.name',
                  value: { stringValue: 'frontend' },
                },
                {
                  key: 'url.full',
                  value: { stringValue: 'http://10.1.217.83:443/ping' },
                },
                {
                  key: 'user_agent',
                  value: { stringValue: 'ELB-HealthChecker/2.0' },
                },
                {
                  key: 'zone',
                  value: { stringValue: '' },
                },
              ],
              droppedAttributesCount: 0,
              events: [],
              droppedEventsCount: 0,
              links: [],
              droppedLinksCount: 0,
              status: {
                code: 0,
                message: '',
              },
              flags: 0,
              traceState: '',
            },
          ],
        },
      ],
    },
  ],
};




